+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-## 2.0.2 [unreleased]
+## 2.0.2
 
 - Added `retain` option to `reindex`
 - Added support for attributes in highlight tags
 - Fixed potentially silent errors in reindex job
-- Improved format for `boost_by_distance`
+- Improved syntax for `boost_by_distance`
 
 ## 2.0.1